+++ src/biz/bokhorst/xprivacy/XGoogleMapV1.java
@@ -3,10 +3,6 @@ package biz.bokhorst.xprivacy;
 import java.util.ArrayList;
 import java.util.List;
 
-import android.location.Location;
-import android.os.Binder;
-import android.util.Log;
-
 public class XGoogleMapV1 extends XHook {
 	private Methods mMethod;
 
@@ -16,67 +12,42 @@ public class XGoogleMapV1 extends XHook {
 	}
 
 	public String getClassName() {
-		if (mMethod == Methods.getLatitudeE6 || mMethod == Methods.getLongitudeE6)
-			return "com.google.android.maps.GeoPoint";
-		else
-			return "com.google.android.maps.MyLocationOverlay";
+		return "com.google.android.maps.MyLocationOverlay";
 	}
 
-	// @formatter:off
-
-	// int getLatitudeE6()
-	// int getLongitudeE6()
 	// boolean enableMyLocation()
-	// GeoPoint getMyLocation() 
+	// void disableMyLocation()
 	// https://developers.google.com/maps/documentation/android/v1/reference/index
 
-	// @formatter:on
-
 	private enum Methods {
-		getLatitudeE6, getLongitudeE6, enableMyLocation
+		enableMyLocation, disableMyLocation
 	};
 
 	public static List<XHook> getInstances() {
 		List<XHook> listHook = new ArrayList<XHook>();
-		for (Methods method : Methods.values())
-			listHook.add(new XGoogleMapV1(method, PrivacyManager.cLocation));
+		listHook.add(new XGoogleMapV1(Methods.enableMyLocation, PrivacyManager.cLocation));
+		listHook.add(new XGoogleMapV1(Methods.disableMyLocation, null));
 		return listHook;
 	}
 
 	@Override
 	protected void before(XParam param) throws Throwable {
-		if (mMethod == Methods.getLatitudeE6) {
-			// Do nothing
-
-		} else if (mMethod == Methods.getLongitudeE6) {
-			// Do nothing
-
-		} else if (mMethod == Methods.enableMyLocation) {
+		switch (mMethod) {
+		case enableMyLocation:
 			if (isRestricted(param))
 				param.setResult(false);
+			break;
 
-		} else
-			Util.log(this, Log.WARN, "Unknown method=" + param.method.getName());
+		case disableMyLocation:
+			if (isRestricted(param, PrivacyManager.cLocation, "MapV1.enableMyLocation"))
+				if (isRestricted(param))
+					param.setResult(null);
+			break;
+		}
 	}
 
 	@Override
 	protected void after(XParam param) throws Throwable {
-		if (mMethod == Methods.getLatitudeE6) {
-			if (isRestricted(param)) {
-				Location fakeLocation = PrivacyManager.getDefacedLocation(Binder.getCallingUid(), null);
-				param.setResult((int) Math.round(fakeLocation.getLatitude() * 1e6));
-			}
-
-		} else if (mMethod == Methods.getLongitudeE6) {
-			if (isRestricted(param)) {
-				Location fakeLocation = PrivacyManager.getDefacedLocation(Binder.getCallingUid(), null);
-				param.setResult((int) Math.round(fakeLocation.getLongitude() * 1e6));
-			}
-
-		} else if (mMethod == Methods.enableMyLocation) {
-			// Do nothing
-
-		} else
-			Util.log(this, Log.WARN, "Unknown method=" + param.method.getName());
+		// Do nothing
 	}
 }